Spring Integration 예제 - H2 저장 및 조회 테스트
- AI 요청 및 응답처리 스레드가 1개이기 때문에 현재 구조에서는 좌표 요청 속도가 처리 속도보다 빠르면 sendAt의 Interval이 늘어나는 구조임 - 순서 보장을 위해 의도된 처리가 맞음

+++ src/main/java/yoon/jy/practice/repository/IdMemoryRepository.java
@@ -1,14 +1,14 @@
 package yoon.jy.practice.repository;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.stereotype.Component;
 
 @Component
 public class IdMemoryRepository {
 
-  private final AtomicInteger traceId = new AtomicInteger(0);
+  private final AtomicLong traceId = new AtomicLong(0);
 
-  public Integer generateTraceId() {
+  public Long generateTraceId() {
     return traceId.incrementAndGet();
   }
 }
